refactor: split card tables out of models.py, and reset the add forms
Two findings left open by the last review.

models.py had gone from 922 lines to 1,113 -- this branch pushed it over the
1k mark, which is the thing that rule exists to catch. Cards are a
self-contained subdomain (three tables and three enums nothing else declares),
so they move to card_models.py and models.py re-exports them; every caller
keeps saying `models.Card`. The re-export is load-bearing rather than cosmetic:
it is what registers the tables on Base.metadata before Alembic reads it.
Relationships resolve by class name through SQLAlchemy's registry, so nothing
imports back the other way. models.py is 949 lines again, and `alembic check`
reports no new upgrade operations -- the split is metadata-identical.

The add-limit and add-category forms left the previous entry sitting in their
inputs after a successful submit, one mis-click from creating it twice. They
now submit through a fetcher and clear once the server has accepted, which also
means the list updates in place and each form can show its own error instead of
throwing it to the page banner.

Verified in the browser: adding "Dining cap / 600" creates it, clears the
fields, and updates the list without a navigation; deleting it still works.
