+++ AGENTS.md
@@ -103,6 +103,7 @@ This document provides a high-level overview and instructions for AI agents work
     - Deleting a card category still referenced by transactions is a **409 with an explanation**, the same treatment deleting a category used by a recurring rule gets. Deleting a *limit* leaves its categories in place and merely unmetered (`ON DELETE SET NULL`) — removing a cap must not delete the user's taxonomy.
     - **Changing a transaction's account clears `card_category_id`.** A card category belongs to one card, so moving the row elsewhere makes the pick meaningless; it is cleared rather than left dangling into another card's taxonomy. An explicit pick in the same request still wins and is validated against the account being moved to.
     - Scoping runs **entirely through the card's financial account**, which already carries the household and the private-ownership rule. `_visible_card` is the single gate; the household list endpoint applies the ownership predicate in the query, not after it.
+    - Card tables live in `backend/src/card_models.py`, not `models.py`, and are re-exported from it so callers keep saying `models.Card`. That re-export is load-bearing twice over: it is what registers them on `Base.metadata` before Alembic reads it, and it is why relationships resolve — they are matched by class name through SQLAlchemy's registry, so there is no import back the other way and no cycle. `alembic check` reports no drift across the split.
     - Web fetches the card list and the picker's headroom through **resource routes on demand**, not the Transactions page loader — most accounts are not cards. iOS and Android are not built yet; the parity modules will be `ios/.../Support/Cards.swift` and `android/.../logic/Cards.kt` against `frontend/src/lib/cards.ts`.
 
 - **Multi-Currency Reporting**: The system standardizes all financial reporting (Dashboard, Portfolio, Net Worth) to the household's `base_currency`.
